+++ esm/lib/index.js
@@ -1,0 +1,30 @@
+// eslint-disable-next-line import/extensions
+import pkg from '../../commonjs/lib/index.js';
+
+export const {
+  connector,
+  error,
+  message,
+  metamodel,
+  util,
+  intersection,
+  caching,
+  query,
+  partialupdate,
+  model,
+  EntityManagerFactory,
+  ConnectData,
+  EntityManager,
+  Acl,
+  GeoPoint,
+  baqend,
+  db,
+  connect,
+  configure,
+  Set,
+  Map,
+  List,
+} = pkg;
+
+// eslint-disable-next-line import/no-default-export
+export default db;