+++ lib/util/uuid.ts
@@ -1,7 +1,8 @@
 import { v4 } from 'uuid';
 
 const uuid = v4 as uuid;
-// eslint-disable-next-line @typescript-eslint/naming-convention
+
+// eslint-disable-next-line @typescript-eslint/naming-convention,@typescript-eslint/no-redeclare
 interface uuid {
   /**
      * Generates a new Universally Unique Identifier (UUID) version 4.